- reuse data fields from to-be-discarded levels in checkpoint instead of   creating new duplicates and discarding the old ones
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@3808 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ app/undo.h
@@ -28,7 +28,7 @@ G_BEGIN_DECLS
 gulong          gwy_app_undo_checkpoint           (GwyContainer *data,
                                                    ...);
 gulong          gwy_app_undo_checkpointv          (GwyContainer *data,
-                                                   gsize n,
+                                                   guint n,
                                                    const gchar **keys);
 void            gwy_app_undo_undo                 (void);
 void            gwy_app_undo_undo_window          (GwyDataWindow *data_window);
